requirements.txt
flask==2.1.0
psycopg2-binary==2.9.3
gunicorn==20.1.0
pytest==7.1.1
requests==2.27.1